+++ Event.cpp
@@ -1,6 +1,6 @@
 #include "Event.h"
 
-Event::Event(float t, int ty){
+Event::Event(unsigned long t, int ty){
   time = t;
   eType = ty;
 }
@@ -9,7 +9,7 @@ int Event::getType(){
   return eType;
 }
 
-float Event::getTime(){
+unsigned long Event::getTime(){
   return time;
 }
